requirements.txt
anyio==3.6.1
boto3==1.24.36
botocore==1.27.36
click==8.1.3
fastapi==0.79.0
ffmpeg==1.4
h11==0.13.0
httptools==0.4.0
idna==3.3
jmespath==1.0.1
pydantic==1.9.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytube==12.1.3
PyYAML==6.0
s3transfer==0.6.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing-extensions==4.3.0
urllib3
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.16.0
websockets==10.3
mutagen==1.45.1
yt-dlp==2023.11.16